./utils: fparse() use 'printf'
Originally, fparse was meant to use printf, but that never happened due
to echo being faster than printf in most cases. Now, using echo seems
like a bad idea because it interprets too much input, e.g. '-c', '\n',
etc.

Even though fparse is meant to be used on files containing "normal
characters", it never hurts for fparse to be able to interpret
any file correctly. On a side note, benchmarking printf and echo shows
that printf can also be the faster utility when interpreting files that
contain more special than normal characters.

Ultimately, using echo to parse an unknown file is a security risk. The
speed difference between both of the utilities is minimal, and both of
them can be the faster one on specific input. With that said, printf is
the better suited tool for fparse.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>