./utils: fparse() set IFS to an empty value
If IFS is not empty, leading and trailing whitespaces will not be
included while using read to parse a file. Recently, fparse() was
changed to use printf for maximum corectness. Since then, our goal is to
make fparse() parse a file with maximum accuracy and corectness.

So, to achieve absolute accuracy, set IFS so that fparse() can parse a
file with exact interpretation, i.e. one-to-one copy. Although this
changes nothing for us, it can be useful in the future; e.g. store a
file in a variable and back, manipulate its content by stripping lines,
etc.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>