spatch: Disable shellcheck warnings (SC2030 and SC2031)
ShellCheck is now totally wrong with these warnings, i.e. it is not
parsing the syntax correctly. It thinks some variables are modified
locally, but they are not - they are global and modified globally.

The code is 100% POSIX and correct, also tested, so this can only be
ShellCheck's own issue. Although it does not have a standard (or I am
not aware of one), the syntax used should be fairly easily parseable and
readable, both by human and by ShellCheck.

TL;DR - false positive.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>