./utils: ltr_substr0() refactor/fix
To sum it up:
- Improved whitespace stripping.
- Improved expansion verification.
- Improved expansion to be as safe as possible across all shells.
- Improved documentation to be as accurate as possible.
- Syntax changes, fixes, and improvements.

The expansion verification part took time to make, improve, verify, and
test... simply said, it is somewhat hard to grasp the code and why it is
what it is. As such, below you can find resources to help you learn more
and "battle-test" the function for correctness in case changes or
improvements have to be made.

Tests: https://raw.githubusercontent.com/mscalindt/top-secret/root/2/1
Ruleset: https://raw.githubusercontent.com/mscalindt/top-secret/root/2/2

Signed-off-by: Dimitar Yurukov <mscalindt@gmail.com>