./utils: Reorder specific simple options
Some options, such as '--no-color', affect the utility behavior
globally. At the same time, there are options that disregard other
options; if an option that affects the utility behavior is specified
after such an option -- it would not activate.

To fix that, arguments such as '--no-color' will take priority over
other positional parameters by reordering them to be first. Currently,
only simple options are supported; at a later stage and when needed,
the utilities can be expanded to support reordering of complex options.

Signed-off-by: Dimitar Yurukov <mscalindt@gmail.com>