./utils: args_ro() fix arg pattern match
Prevent false positives. Using whitespace we can be sure the first and
last character of "_opt" variable are the single quotes delimiting the
arg; without whitespace, we cannot be sure the same quotes are not an
escaped single quote within single quotes (''\'''; '\'') as the escaping
structure always begins and ends with a single quote character.

Signed-off-by: Dimitar Yurukov <mscalindt@gmail.com>