./utils: args_ro() handle grouped short option
Note that a short option scattered among a group of short options is
impossible to be reordered unless it is the starting point of the group,
as it is not known if the short options before the one we reorder are
complex options, i.e. options with an argument.

The problem we face in the above paragraph cannot be solved without
external means, because the function itself does not know which options
are complex (unless we start utilizing our polymorphism ability).
Nevertheless, this problem indicates something much problematic --
relying on reordered positional parameters to illustrate and convey the
behavior of the utility.

Therefore, we learned that, from one point of view we are utilizing an
unstandardized feature and from another point of view the behavior of
the utility is not well-defined and standardized. The timeline of events
has put us in a dilemma -- shall we continue extending the feature, or
shall we spend time documenting and standardizing the command-line of
utilities that require such bizzare feature?

Signed-off-by: Dimitar Yurukov <mscalindt@gmail.com>