./utils: args_ro() use main stack variable names
This function is not meant to be reusable; it runs in the main stack --
the main code path -- of the program.

Functions that are not meant to be reusable, e.g. functions that are
part of the main code path, shall use variables that do not start with
an underscore character, as variables with underscore character are
reserved for the sub stack (reusable) functions.

The above rule is for a reason: it allows reusable/helper functions to
be ran in the main stack without interfering with the variables of the
main stack -- without using a subshell/fork.

Signed-off-by: Dimitar Yurukov <mscalindt@gmail.com>