./utils: args_ro() support a literal '--' string
args_ro() relies on a literal '--' positional parameter to specify the
end of strings to reorder; the drawback being that a literal '--' string
cannot be reordered. This patch fixes that drawback using a dynamic
approach to detect the accurate amount of strings specified, therefore
eliminating the need of a delimiter such as '--'.

Signed-off-by: Dimitar Yurukov <mscalindt@gmail.com>