./utils: Remove args_ro()
Deprecate the function in favor of flag-only options. Even though
args_ro() was heavily polished and improved since its existence, its
core purpose remained iffy at best. In its current state, the function
cannot correctly determine whether the reorder of an argument will not
incorrectly affect the behavior of the utility it is used by.

Therefore, flag-only options are the way ahead. They will be implemented
in a separate patch.

Signed-off-by: Dimitar Yurukov <mscalindt@gmail.com>