admerge: Spit explicit output on branch error
The command that identifies the current branch shall never fail unless:
(1) git version is too old (i.e. the command is incompatible);
(2) the HEAD is detached (i.e. HEAD is not pointing to a branch).

Therefore, if admerge cannot get the current branch, the HEAD is
probably detached; in such state, the concept of "our branch" is
invalid. Indicate the situation accordingly in the error message.

NOTE: admerge has no means of knowing when the first point is true, at
least not directly, which is why it is not accounted for; admerge is a
wrapper, not a helper of some sort. Compatibility with old versions is
accounted for as much as git allows.

Signed-off-by: Dimitar Yurukov <mscalindt@gmail.com>